+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Changed
+
+- `bill`: `Line` and `SubLine` now normalize negative item prices to quantity.
+- `org`: `Item` requires zero or positive price.
+
 ## [v0.305.1] - 2026-01-09
 
 ### Fixed
@@ -15,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ## [v0.305.0] - 2026-01-08
 
 ### Added
+
 - `ar-arca-v4`: Added initial Argentina ARCA addon.
 - `eu-en16931-v2017`: Added check for negative prices.
 - `br-nfse-v1`: Add extensions for IBS/CBS tax determination (RTC)